wai-aria/subtree/basic.tentative.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/wai-aria/scripts/aria-utils.js"></script>

<div id="tablist" role="tablist" aria-label="tablist">
  <div id="tab1" role="tab" aria-label="tab1"></div>
  <div id="tab2" role="tab" aria-label="tab2"></div>
</div>

<script>

promise_test(async t => {
  const tablist = await test_driver.get_element_accessible_properties(document.getElementById("tablist"));
  assert_equals(tablist.children.length, 2);
  const tab1 = await test_driver.get_accessible_properties(tablist.children[0]);
  assert_equals(tab1.id, tablist.children[0]);
  assert_equals(tab1.parent, tablist.id);
  assert_equals(tab1.role, "tab");
  assert_equals(tab1.label, "tab1");
  const tab2 = await test_driver.get_accessible_properties(tablist.children[1]);
  assert_equals(tab2.id, tablist.children[1]);
  assert_equals(tab2.parent, tablist.id);
  assert_equals(tab2.role, "tab");
  assert_equals(tab2.label, "tab2");
}, "tablist explicit tree");

AriaUtils.verifyAccessibilitySubtree("#tablist", {
  role: "tablist",
  label: "tablist",
  children: [
    { role: "tab", label: "tab1", children: [] },
    { role: "tab", label: "tab2", children: [] },
  ],
});

</script>
